refactor(manoeuvre): #7 — éclate algo.py en package algo/ en couches
L'ancien module « god » algo.py (~2450 lignes, 45 symboles, 4 points d'entrée
mêlant placement, séquencement, vérification et helpers bas niveau) devient un
package en couches à dépendances strictement descendantes (sans cycle) :

    _constants / results → graph_ops → placement / verification
                         → sequencing → targets

- results       — Manoeuvre, ResultatManoeuvres
- graph_ops     — index O(1), _is_open/_set_switch, chemins SA, couplers inter-SJB
- placement     — _placement_automatique (+ best-effort/glouton), _main_busbar_sjb
- verification  — _rejeu_securite, sectionneurs_sous_charge_par_manoeuvre, _optimiser_sequence
- sequencing    — determiner_manoeuvres_avec_sections, ré-aiguillages
- targets       — determiner_topo_complete_cible / _cible_detaillee (smooth/aggressive/multibarres)

algo/__init__.py réexporte INTÉGRALEMENT la surface de l'ancien module (publics
ET privés, ré-exports explicites `X as X`) : manoeuvre.algo.X et
`from ...algo import X` restent inchangés pour les 23 sites d'import — zéro
changement côté tests/scripts/IHM.

Iso-comportement : chaque symbole déplacé verbatim (corps identiques vérifiés
par comparaison d'AST symbole-par-symbole). 646 tests verts, goldens
déterministes sur ≥ 2 graines, couverture ~92 % inchangée, ruff + interrogate OK.
